working on checking and getting participants
