+++ NFACT/NFACT_PP/nfactpp_args.py
@@ -187,7 +187,7 @@ def nfact_pp_example_usage() -> str:
 
 Example Usage:
     {col['purple']}Seed surface mode:{col['reset']}
-           python3 -m NFACT_PP --study_folder /home/mr_robot/subjects 
+           nfact_pp --study_folder /home/mr_robot/subjects 
                --list /home/mr_robot/sub_list  
                --bpx_path Diffusion.bedpostX 
                --seeds L.white.32k_fs_LR.surf.gii R.white.32k_fs_LR.surf.gii 
@@ -197,7 +197,7 @@ Example Usage:
                --gpu --n_cores 3 
            \n
     {col['pink']}Volume surface mode:{col['reset']}
-            python3 -m NFACT_PP --study_folder /home/mr_robot/subjects 
+            nfact_pp --study_folder /home/mr_robot/subjects 
                 --list /home/mr_robot/sub_list  
                 --bpx_path Diffusion.bedpostX 
                 --seeds L.white.nii.gz R.white.nii.gz 
@@ -207,7 +207,7 @@ Example Usage:
                 --gpu --n_cores 3 
         \n
     {col['darker_pink']}HCP mode:{col['reset']}
-        python3 -m NFACT_PP --hcp_stream
+        nfact_pp --hcp_stream
             --study_folder /home/mr_robot/subjects  
             --list /home/mr_robot/for_axon/nfact_dev/sub_list  
             --image_standard_space $FSLDIR/data/standard/MNI152_T1_2mm_brain.nii.gz
